LayoutTests/fast/css/getComputedStyle/getComputedStyle-text-decoration.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
function testElementStyle(value, lineValue)
{
    if (!lineValue)
        lineValue = value;

    shouldBe("e.style.textDecoration", "'" + value + "'");
    shouldBe("e.style.getPropertyValue('text-decoration')", "'" + value + "'");
    shouldBeNull("e.style.getPropertyCSSValue('text-decoration')");

    shouldBe("e.style.textDecorationLine", "'" + lineValue + "'");
    shouldBe("e.style.getPropertyValue('text-decoration-line')", "'" + lineValue + "'");
    if (lineValue)
        shouldBe("e.style.getPropertyCSSValue('text-decoration-line').cssText", "'" + lineValue + "'");
    else
        shouldBeNull("e.style.getPropertyCSSValue('text-decoration-line')");
}

function testComputedStyle(value, lineValue)
{
    if (!lineValue)
        lineValue = value;

    computedStyle = window.getComputedStyle(e, null);

    shouldBe("computedStyle.textDecoration", "'" + value + "'");
    shouldBe("computedStyle.getPropertyValue('text-decoration')", "'" + value + "'");
    shouldBe("computedStyle.getPropertyCSSValue('text-decoration').cssText", "'" + value + "'");

    shouldBe("computedStyle.textDecorationLine", "'" + lineValue + "'");
    shouldBe("computedStyle.getPropertyValue('text-decoration-line')", "'" + lineValue + "'");
    if (lineValue)
        shouldBe("computedStyle.getPropertyCSSValue('text-decoration-line').cssText", "'" + lineValue + "'");
    else
        shouldBeNull("computedStyle.getPropertyCSSValue('text-decoration-line')");
}

description("Test to make sure text-decoration property returns values properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<div id="test">hello world</div>';
debug("Initial value:");
e = document.getElementById('test');
testElementStyle('');
testComputedStyle("rgb(0, 0, 0)", "none");
debug('');

debug("Initial value (explicit):");
e.style.textDecoration = 'initial';
testElementStyle("initial");
testComputedStyle("rgb(0, 0, 0)", "none");
debug('');

debug("Value 'none':");
e.style.textDecoration = 'none';
testElementStyle("none");
testComputedStyle("rgb(0, 0, 0)", "none");
debug('');

debug("Value 'underline':");
e.style.textDecoration = 'underline';
testElementStyle("underline", "underline");
testComputedStyle("underline rgb(0, 0, 0)", "underline");
debug('');

debug("Value 'overline':");
e.style.textDecoration = 'overline';
testElementStyle("overline");
testComputedStyle("overline rgb(0, 0, 0)", "overline");
debug('');

debug("Value 'line-through':");
e.style.textDecoration = 'line-through';
testElementStyle("line-through");
testComputedStyle("line-through rgb(0, 0, 0)", "line-through");
debug('');

debug("Value 'underline overline line-through':");
e.style.textDecoration = 'underline overline line-through';
testElementStyle("underline overline line-through");
testComputedStyle("underline overline line-through rgb(0, 0, 0)", "underline overline line-through");
debug('');

debug("Value 'blink' (valid but ignored):");
e.style.textDecoration = 'blink';
testElementStyle("blink");
testComputedStyle("none rgb(0, 0, 0)", "none");
debug('');

debug("Value '':");
e.style.textDecoration = '';
testElementStyle('');
testComputedStyle("rgb(0, 0, 0)", "none");
debug('');

testContainer.innerHTML = '<div id="test-parent" style="text-decoration: underline;">hello <span id="test-ancestor" style="text-decoration: inherit;">world</span></div>';
debug("Parent gets 'underline' value:");
e = document.getElementById('test-parent');
testElementStyle("underline");
testComputedStyle("underline rgb(0, 0, 0)", "underline");
debug('');

debug("Ancestor should explicitly inherit value from parent when 'inherit' value is used:");
e = document.getElementById('test-ancestor');
testElementStyle("inherit");
testComputedStyle("underline rgb(0, 0, 0)", "underline");
debug('');

debug("Ancestor should not implicitly inherit value from parent (i.e. when value is void):");
e.style.textDecoration = '';
testElementStyle('');
testComputedStyle("rgb(0, 0, 0)", "none");
debug('');

document.body.removeChild(testContainer);
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
